Add support of multiple renderers of the same type
If different audio renderers extensions connected to the ExoPlayer (ffmpeg extension, for example) and the video file contains audio tracks of different types handled by different renderers (for example aac and ac-3) there will be multiple renderer tracks in the MappedTrackInfo. Only one renderer of the same type may work properly at the time, so if selecting track in one renderer another renderer should be disabled to avoid playback errors.

This commits tries to handle multiple renderers automatically in the transparent way. Resulting API is untouched so it should be possible to work with multiple renderers of the same type in the same way as with a single renderer.